requirements.txt
Flask==3.0.0
Flask-OAuthlib
requests
python-dotenv